move players

+++ src/Model/Move.cs
@@ -5,9 +5,9 @@ namespace Model
     internal class Move
     {
         public Guid PlayerId { get; }
-        public float Position { get; }
+        public double Position { get; }
 
-        public Move(Guid playerId, float position)
+        public Move(Guid playerId, double position)
         {
             PlayerId = playerId;
             Position = position;
